DO NOT COMMIT - Moved parser and lexer into their own files - updated gitignore for arch venv directories
