+++ src/umleditor/mvc_view/__init__.py
@@ -1,0 +1,1 @@
+from .cli_lexer import lex_input as lex
